Backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-multipart==0.0.20
pdfplumber==0.11.7
requests==2.31.0
python-dotenv==1.1.1
